tidied up the telemetry module
