+++ framework/telemetry/telemetry.qrc
